feat testing

+++ overlap task part 1/groups/src/groups/groups.repository.ts
@@ -6,26 +6,34 @@ export class GroupRepository {
     const a = GroupModel.findById(mainGroupId) as unknown as Group;
     return a;
   }
+
+  static gelAllGroups(): Promise<Group[]> {
+    return GroupModel.find().exec();
+  }
+
   static create(newGroup: Group): Promise<Group> {
     return GroupModel.create(newGroup);
   }
 
-  static findGroupByID(groupId: string): Promise<Group[]> {
-    return GroupModel.find({ groupId }).exec();
+  static findGroupByID(groupId: string): Promise<Group | null> {
+    return GroupModel.findOne({ groupId }).exec();
   }
 
-  static updateGroupByID(groupId: string, postData: Partial<Group>): Promise<Group | null> {
-    return GroupModel.findByIdAndUpdate(groupId, postData, { upsert: true }).exec();
+  static async updateGroupByID(groupId: string, postData: Partial<Group>): Promise<Group | null> {
+    const token = await GroupModel.findByIdAndUpdate(groupId, postData, { upsert: true, new: true }).exec();
+    return token;
   }
 
   static async deleteGroupByID(groupId: string): Promise<Group | null> {
-    const group = await GroupModel.findById(groupId).exec();
-    await GroupRepository.deleteSubgroups(group);
+    const group = await GroupModel.findById(groupId).exec();  
+    if(group?.subgroups){
+      await GroupRepository.deleteSubgroups(group);
+    }
     return GroupModel.findByIdAndDelete(groupId).exec();
   }
 
   static async deleteSubgroups(group: any): Promise<void> {
-    const subgroups = group?.subgroups;
+    const subgroups = group.subgroups;
     for (const subgroupId of subgroups) {
       await GroupModel.findByIdAndDelete(subgroupId).exec();
     }
@@ -42,6 +50,11 @@ export class GroupRepository {
   static showGroupHierarchy(id: string, groups: Group[]): Promise<Group[]> {
     return GroupModel.find({ id }, groups).exec();
   }
+
+  static findPerson(personName: string): Promise<Group[]> {
+    return GroupModel.find({ personName }).exec();
+  }
+
 }
